Verified

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: tamigore <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 14:44:27 by tamigore          #+#    #+#             */
-/*   Updated: 2018/11/14 19:29:22 by tamigore         ###   ########.fr       */
+/*   Updated: 2018/11/19 16:20:38 by tamigore         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	size_t i;
+	int	i;
 
-	i = len - 1;
+	i = (int)len - 1;
 	if ((unsigned char *)dst < (unsigned char *)src)
 		ft_memcpy(dst, src, len);
 	else
 	{
-		while ((int)i >= 0)
+		while (i >= 0)
 		{
 			((unsigned char *)dst)[i] = ((unsigned char *)src)[i];
 			i--;
