+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: tamigore <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/09 09:58:06 by tamigore          #+#    #+#             */
-/*   Updated: 2018/11/12 13:20:16 by tamigore         ###   ########.fr       */
+/*   Updated: 2018/11/19 16:14:09 by tamigore         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = ft_strlen(s1);
 	j = 0;
-	while (s2[j] && j < (int)n)
+	while (s2[j] && j < n)
 		s1[i++] = s2[j++];
 	s1[i] = '\0';
 	return (s1);
